stm32_metapac::otg::regs

Struct GccfgV3

#[repr(transparent)]
pub struct GccfgV3(pub u32);

OTG general core configuration register.

Tuple Fields
0: u32
Implementations

impl GccfgV3

pub const fn chgdet(&self) -> bool

Charger detection, result of the current mode (primary or secondary).

pub fn set_chgdet(&mut self, val: bool)

Charger detection, result of the current mode (primary or secondary).

pub const fn fsvplus(&self) -> bool

Single-Ended DP indicator This bit gives the voltage level on DP (also result of the comparison with VLGC threshold as defined in BC v1.2 standard).

pub fn set_fsvplus(&mut self, val: bool)

Single-Ended DP indicator This bit gives the voltage level on DP (also result of the comparison with VLGC threshold as defined in BC v1.2 standard).

pub const fn fsvminus(&self) -> bool

Single-Ended DM indicator This bit gives the voltage level on DM (also result of the comparison with VLGC threshold as defined in BC v1.2 standard).

pub fn set_fsvminus(&mut self, val: bool)

Single-Ended DM indicator This bit gives the voltage level on DM (also result of the comparison with VLGC threshold as defined in BC v1.2 standard).

pub const fn sessvld(&self) -> bool

VBUS session indicator Indicates if VBUS is above VBUS session threshold.

pub fn set_sessvld(&mut self, val: bool)

VBUS session indicator Indicates if VBUS is above VBUS session threshold.

pub const fn hcdpen(&self) -> bool

Host CDP behavior enable.

pub fn set_hcdpen(&mut self, val: bool)

Host CDP behavior enable.

pub const fn hcdpdeten(&self) -> bool

Host CDP port voltage detector enable on DP.

pub fn set_hcdpdeten(&mut self, val: bool)

Host CDP port voltage detector enable on DP.

pub const fn hvdmsrcen(&self) -> bool

Host CDP port Voltage source enable on DM.

pub fn set_hvdmsrcen(&mut self, val: bool)

Host CDP port Voltage source enable on DM.

pub const fn dcden(&self) -> bool

Data Contact Detection enable.

pub fn set_dcden(&mut self, val: bool)

Data Contact Detection enable.

pub const fn pden(&self) -> bool

Primary detection enable.

pub fn set_pden(&mut self, val: bool)

Primary detection enable.

pub const fn vbden(&self) -> bool

VBUS detection enable Enables VBUS Sensing Comparators in order to detect VBUS presence and/or perform OTG operation.

pub fn set_vbden(&mut self, val: bool)

VBUS detection enable Enables VBUS Sensing Comparators in order to detect VBUS presence and/or perform OTG operation.

pub const fn sden(&self) -> bool

Secondary detection enable.

pub fn set_sden(&mut self, val: bool)

Secondary detection enable.

pub const fn vbvaloval(&self) -> bool

Software override value of the VBUS B-session detection.

pub fn set_vbvaloval(&mut self, val: bool)

Software override value of the VBUS B-session detection.

pub const fn vbvaloven(&self) -> bool

Enables a software override of the VBUS B-session detection.

pub fn set_vbvaloven(&mut self, val: bool)

Enables a software override of the VBUS B-session detection.

pub const fn forcehostpd(&self) -> bool

Force host mode pull-downs If the ID pin functions are enabled, the host mode pull-downs on DP and DM activate automatically. However, whenever that is not the case, yet host mode is required, this bit must be used to force the pull-downs active.

pub fn set_forcehostpd(&mut self, val: bool)

Force host mode pull-downs If the ID pin functions are enabled, the host mode pull-downs on DP and DM activate automatically. However, whenever that is not the case, yet host mode is required, this bit must be used to force the pull-downs active.