stm32_metapac::pwr::regs

Struct Cr3

#[repr(transparent)]
pub struct Cr3(pub u32);

Reset only by POR only, not reset by wakeup from Standby mode and RESET pad. The lower byte of this register is written once after POR and shall be written before changing VOS level or ck_sys clock frequency. No limitation applies to the upper bytes.Programming data corresponding to an invalid combination of SDLEVEL, SDEXTHP, SDEN, LDOEN and BYPASS bits (see Table9) will be ignored: data will not be written, the written-once mechanism will lock the register and any further write access will be ignored. The default supply configuration will be kept and the ACTVOSRDY bit in PWR control status register 1 (PWR_CSR1) will go on indicating invalid voltage levels. The system shall be power cycled before writing a new value.

Tuple Fields
0: u32
Implementations

impl Cr3

pub const fn bypass(&self) -> bool

Power management unit bypass

pub fn set_bypass(&mut self, val: bool)

Power management unit bypass

pub const fn ldoen(&self) -> bool

Low drop-out regulator enable

pub fn set_ldoen(&mut self, val: bool)

Low drop-out regulator enable

pub const fn sden(&self) -> bool

SD converter Enable

pub fn set_sden(&mut self, val: bool)

SD converter Enable

pub const fn sdexthp(&self) -> bool

Step-down converter forced ON and in High Power MR mode

pub fn set_sdexthp(&mut self, val: bool)

Step-down converter forced ON and in High Power MR mode

pub const fn sdlevel(&self) -> u8

Step-down converter voltage output level selection

pub fn set_sdlevel(&mut self, val: u8)

Step-down converter voltage output level selection

pub const fn vbe(&self) -> bool

VBAT charging enable

pub fn set_vbe(&mut self, val: bool)

VBAT charging enable

pub const fn vbrs(&self) -> bool

VBAT charging resistor selection

pub fn set_vbrs(&mut self, val: bool)

VBAT charging resistor selection

pub const fn sdextrdy(&self) -> bool

SMPS step-down converter external supply ready

pub fn set_sdextrdy(&mut self, val: bool)

SMPS step-down converter external supply ready

pub const fn usb33den(&self) -> bool

VDD33USB voltage level detector enable.

pub fn set_usb33den(&mut self, val: bool)

VDD33USB voltage level detector enable.

pub const fn usbregen(&self) -> bool

USB regulator enable.

pub fn set_usbregen(&mut self, val: bool)

USB regulator enable.

pub const fn usb33rdy(&self) -> bool

USB supply ready.

pub fn set_usb33rdy(&mut self, val: bool)

USB supply ready.