Struct stm32_metapac::pwr::regs::Wkupepr

#[repr(transparent)]
pub struct Wkupepr(pub u32);

Reset only by system reset, not reset by wakeup from Standby mode

Tuple Fields
0: u32
Implementations

impl Wkupepr

pub const fn wkupen(&self, n: usize) -> bool

Enable Wakeup Pin WKUPn+1 Each bit is set and cleared by software. Note: An additional wakeup event is detected if WKUPn+1 pin is enabled (by setting the WKUPENn+1 bit) when WKUPn+1 pin level is already high when WKUPPn+1 selects rising edge, or low when WKUPPn+1 selects falling edge.

pub fn set_wkupen(&mut self, n: usize, val: bool)

Enable Wakeup Pin WKUPn+1 Each bit is set and cleared by software. Note: An additional wakeup event is detected if WKUPn+1 pin is enabled (by setting the WKUPENn+1 bit) when WKUPn+1 pin level is already high when WKUPPn+1 selects rising edge, or low when WKUPPn+1 selects falling edge.

pub const fn wkupp(&self, n: usize) -> bool

Wakeup pin polarity bit for WKUPn-7 These bits define the polarity used for event detection on WKUPn-7 external wakeup pin.

pub fn set_wkupp(&mut self, n: usize, val: bool)

Wakeup pin polarity bit for WKUPn-7 These bits define the polarity used for event detection on WKUPn-7 external wakeup pin.

pub const fn wkuppupd(&self, n: usize) -> u8

Wakeup pin pull configuration

pub fn set_wkuppupd(&mut self, n: usize, val: u8)

Wakeup pin pull configuration