stm32_metapac::sdmmc::regs

Struct Icr

#[repr(transparent)]
pub struct Icr(pub u32);

The SDMMC_ICR register is a write-only register. Writing a bit with 1 clears the corresponding bit in the SDMMC_STAR status register.

Tuple Fields
0: u32
Implementations

impl Icr

pub const fn ccrcfailc(&self) -> bool

CCRCFAIL flag clear bit Set by software to clear the CCRCFAIL flag.

pub fn set_ccrcfailc(&mut self, val: bool)

CCRCFAIL flag clear bit Set by software to clear the CCRCFAIL flag.

pub const fn dcrcfailc(&self) -> bool

DCRCFAIL flag clear bit Set by software to clear the DCRCFAIL flag.

pub fn set_dcrcfailc(&mut self, val: bool)

DCRCFAIL flag clear bit Set by software to clear the DCRCFAIL flag.

pub const fn ctimeoutc(&self) -> bool

CTIMEOUT flag clear bit Set by software to clear the CTIMEOUT flag.

pub fn set_ctimeoutc(&mut self, val: bool)

CTIMEOUT flag clear bit Set by software to clear the CTIMEOUT flag.

pub const fn dtimeoutc(&self) -> bool

DTIMEOUT flag clear bit Set by software to clear the DTIMEOUT flag.

pub fn set_dtimeoutc(&mut self, val: bool)

DTIMEOUT flag clear bit Set by software to clear the DTIMEOUT flag.

pub const fn txunderrc(&self) -> bool

TXUNDERR flag clear bit Set by software to clear TXUNDERR flag.

pub fn set_txunderrc(&mut self, val: bool)

TXUNDERR flag clear bit Set by software to clear TXUNDERR flag.

pub const fn rxoverrc(&self) -> bool

RXOVERR flag clear bit Set by software to clear the RXOVERR flag.

pub fn set_rxoverrc(&mut self, val: bool)

RXOVERR flag clear bit Set by software to clear the RXOVERR flag.

pub const fn cmdrendc(&self) -> bool

CMDREND flag clear bit Set by software to clear the CMDREND flag.

pub fn set_cmdrendc(&mut self, val: bool)

CMDREND flag clear bit Set by software to clear the CMDREND flag.

pub const fn cmdsentc(&self) -> bool

CMDSENT flag clear bit Set by software to clear the CMDSENT flag.

pub fn set_cmdsentc(&mut self, val: bool)

CMDSENT flag clear bit Set by software to clear the CMDSENT flag.

pub const fn dataendc(&self) -> bool

DATAEND flag clear bit Set by software to clear the DATAEND flag.

pub fn set_dataendc(&mut self, val: bool)

DATAEND flag clear bit Set by software to clear the DATAEND flag.

pub const fn dholdc(&self) -> bool

DHOLD flag clear bit Set by software to clear the DHOLD flag.

pub fn set_dholdc(&mut self, val: bool)

DHOLD flag clear bit Set by software to clear the DHOLD flag.

pub const fn dbckendc(&self) -> bool

DBCKEND flag clear bit Set by software to clear the DBCKEND flag.

pub fn set_dbckendc(&mut self, val: bool)

DBCKEND flag clear bit Set by software to clear the DBCKEND flag.

pub const fn dabortc(&self) -> bool

DABORT flag clear bit Set by software to clear the DABORT flag.

pub fn set_dabortc(&mut self, val: bool)

DABORT flag clear bit Set by software to clear the DABORT flag.

pub const fn busyd0endc(&self) -> bool

BUSYD0END flag clear bit Set by software to clear the BUSYD0END flag.

pub fn set_busyd0endc(&mut self, val: bool)

BUSYD0END flag clear bit Set by software to clear the BUSYD0END flag.

pub const fn sdioitc(&self) -> bool

SDIOIT flag clear bit Set by software to clear the SDIOIT flag.

pub fn set_sdioitc(&mut self, val: bool)

SDIOIT flag clear bit Set by software to clear the SDIOIT flag.

pub const fn ackfailc(&self) -> bool

ACKFAIL flag clear bit Set by software to clear the ACKFAIL flag.

pub fn set_ackfailc(&mut self, val: bool)

ACKFAIL flag clear bit Set by software to clear the ACKFAIL flag.

pub const fn acktimeoutc(&self) -> bool

ACKTIMEOUT flag clear bit Set by software to clear the ACKTIMEOUT flag.

pub fn set_acktimeoutc(&mut self, val: bool)

ACKTIMEOUT flag clear bit Set by software to clear the ACKTIMEOUT flag.

pub const fn vswendc(&self) -> bool

VSWEND flag clear bit Set by software to clear the VSWEND flag.

pub fn set_vswendc(&mut self, val: bool)

VSWEND flag clear bit Set by software to clear the VSWEND flag.

pub const fn ckstopc(&self) -> bool

CKSTOP flag clear bit Set by software to clear the CKSTOP flag.

pub fn set_ckstopc(&mut self, val: bool)

CKSTOP flag clear bit Set by software to clear the CKSTOP flag.

pub const fn idmatec(&self) -> bool

IDMA transfer error clear bit Set by software to clear the IDMATE flag.

pub fn set_idmatec(&mut self, val: bool)

IDMA transfer error clear bit Set by software to clear the IDMATE flag.

pub const fn idmabtcc(&self) -> bool

IDMA buffer transfer complete clear bit Set by software to clear the IDMABTC flag.

pub fn set_idmabtcc(&mut self, val: bool)

IDMA buffer transfer complete clear bit Set by software to clear the IDMABTC flag.