Struct stm32_metapac::sdmmc::regs::Maskr

#[repr(transparent)]
pub struct Maskr(pub u32);

The interrupt mask register determines which status flags generate an interrupt request by setting the corresponding bit to 1.

Tuple Fields
0: u32
Implementations

impl Maskr

pub const fn ccrcfailie(&self) -> bool

Command CRC fail interrupt enable Set and cleared by software to enable/disable interrupt caused by command CRC failure.

pub fn set_ccrcfailie(&mut self, val: bool)

Command CRC fail interrupt enable Set and cleared by software to enable/disable interrupt caused by command CRC failure.

pub const fn dcrcfailie(&self) -> bool

Data CRC fail interrupt enable Set and cleared by software to enable/disable interrupt caused by data CRC failure.

pub fn set_dcrcfailie(&mut self, val: bool)

Data CRC fail interrupt enable Set and cleared by software to enable/disable interrupt caused by data CRC failure.

pub const fn ctimeoutie(&self) -> bool

Command timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by command timeout.

pub fn set_ctimeoutie(&mut self, val: bool)

Command timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by command timeout.

pub const fn dtimeoutie(&self) -> bool

Data timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by data timeout.

pub fn set_dtimeoutie(&mut self, val: bool)

Data timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by data timeout.

pub const fn txunderrie(&self) -> bool

Tx FIFO underrun error interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO underrun error.

pub fn set_txunderrie(&mut self, val: bool)

Tx FIFO underrun error interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO underrun error.

pub const fn rxoverrie(&self) -> bool

Rx FIFO overrun error interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO overrun error.

pub fn set_rxoverrie(&mut self, val: bool)

Rx FIFO overrun error interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO overrun error.

pub const fn cmdrendie(&self) -> bool

Command response received interrupt enable Set and cleared by software to enable/disable interrupt caused by receiving command response.

pub fn set_cmdrendie(&mut self, val: bool)

Command response received interrupt enable Set and cleared by software to enable/disable interrupt caused by receiving command response.

pub const fn cmdsentie(&self) -> bool

Command sent interrupt enable Set and cleared by software to enable/disable interrupt caused by sending command.

pub fn set_cmdsentie(&mut self, val: bool)

Command sent interrupt enable Set and cleared by software to enable/disable interrupt caused by sending command.

pub const fn dataendie(&self) -> bool

Data end interrupt enable Set and cleared by software to enable/disable interrupt caused by data end.

pub fn set_dataendie(&mut self, val: bool)

Data end interrupt enable Set and cleared by software to enable/disable interrupt caused by data end.

pub const fn dholdie(&self) -> bool

Data hold interrupt enable Set and cleared by software to enable/disable the interrupt generated when sending new data is hold in the DPSM Wait_S state.

pub fn set_dholdie(&mut self, val: bool)

Data hold interrupt enable Set and cleared by software to enable/disable the interrupt generated when sending new data is hold in the DPSM Wait_S state.

pub const fn dbckendie(&self) -> bool

Data block end interrupt enable Set and cleared by software to enable/disable interrupt caused by data block end.

pub fn set_dbckendie(&mut self, val: bool)

Data block end interrupt enable Set and cleared by software to enable/disable interrupt caused by data block end.

pub const fn dabortie(&self) -> bool

Data transfer aborted interrupt enable Set and cleared by software to enable/disable interrupt caused by a data transfer being aborted.

pub fn set_dabortie(&mut self, val: bool)

Data transfer aborted interrupt enable Set and cleared by software to enable/disable interrupt caused by a data transfer being aborted.

pub const fn txfifoheie(&self) -> bool

Tx FIFO half empty interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO half empty.

pub fn set_txfifoheie(&mut self, val: bool)

Tx FIFO half empty interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO half empty.

pub const fn rxfifohfie(&self) -> bool

Rx FIFO half full interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO half full.

pub fn set_rxfifohfie(&mut self, val: bool)

Rx FIFO half full interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO half full.

pub const fn rxfifofie(&self) -> bool

Rx FIFO full interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO full.

pub fn set_rxfifofie(&mut self, val: bool)

Rx FIFO full interrupt enable Set and cleared by software to enable/disable interrupt caused by Rx FIFO full.

pub const fn txfifoeie(&self) -> bool

Tx FIFO empty interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO empty.

pub fn set_txfifoeie(&mut self, val: bool)

Tx FIFO empty interrupt enable Set and cleared by software to enable/disable interrupt caused by Tx FIFO empty.

pub const fn busyd0endie(&self) -> bool

BUSYD0END interrupt enable Set and cleared by software to enable/disable the interrupt generated when SDMMC_D0 signal changes from busy to NOT busy following a CMD response.

pub fn set_busyd0endie(&mut self, val: bool)

BUSYD0END interrupt enable Set and cleared by software to enable/disable the interrupt generated when SDMMC_D0 signal changes from busy to NOT busy following a CMD response.

pub const fn sdioitie(&self) -> bool

SDIO mode interrupt received interrupt enable Set and cleared by software to enable/disable the interrupt generated when receiving the SDIO mode interrupt.

pub fn set_sdioitie(&mut self, val: bool)

SDIO mode interrupt received interrupt enable Set and cleared by software to enable/disable the interrupt generated when receiving the SDIO mode interrupt.

pub const fn ackfailie(&self) -> bool

Acknowledgment Fail interrupt enable Set and cleared by software to enable/disable interrupt caused by acknowledgment Fail.

pub fn set_ackfailie(&mut self, val: bool)

Acknowledgment Fail interrupt enable Set and cleared by software to enable/disable interrupt caused by acknowledgment Fail.

pub const fn acktimeoutie(&self) -> bool

Acknowledgment timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by acknowledgment timeout.

pub fn set_acktimeoutie(&mut self, val: bool)

Acknowledgment timeout interrupt enable Set and cleared by software to enable/disable interrupt caused by acknowledgment timeout.

pub const fn vswendie(&self) -> bool

Voltage switch critical timing section completion interrupt enable Set and cleared by software to enable/disable the interrupt generated when voltage switch critical timing section completion.

pub fn set_vswendie(&mut self, val: bool)

Voltage switch critical timing section completion interrupt enable Set and cleared by software to enable/disable the interrupt generated when voltage switch critical timing section completion.

pub const fn ckstopie(&self) -> bool

Voltage Switch clock stopped interrupt enable Set and cleared by software to enable/disable interrupt caused by Voltage Switch clock stopped.

pub fn set_ckstopie(&mut self, val: bool)

Voltage Switch clock stopped interrupt enable Set and cleared by software to enable/disable interrupt caused by Voltage Switch clock stopped.

pub const fn idmabtcie(&self) -> bool

IDMA buffer transfer complete interrupt enable Set and cleared by software to enable/disable the interrupt generated when the IDMA has transferred all data belonging to a memory buffer.

pub fn set_idmabtcie(&mut self, val: bool)

IDMA buffer transfer complete interrupt enable Set and cleared by software to enable/disable the interrupt generated when the IDMA has transferred all data belonging to a memory buffer.